requirements.txt
discord-interactions  # pipx: alias='discord'
pymongo
requests
colorama
urllib3
python-dateutil
dnspython
python-dotenv
youtube-dl